Use uint32_t in cxx::ratio for better codegen
The avr-gcc literally generates 64bit arithmetic shifts
from divisions if int64_t is used as a denominator. This change helps to
make sure the generated code is more efficient for our small
microcontroller.

+++ src/cxx_ratio.h
@@ -20,11 +20,11 @@ gcd(M m, N n)
 }
 
 
-template<int64_t Num, int64_t Denom = 1>
+template<uint32_t Num, uint32_t Denom = 1>
 struct ratio
 {
-  static constexpr int64_t num = Num / gcd(Num, Denom);
-  static constexpr int64_t den = Denom / gcd(Num, Denom);
+  static constexpr uint32_t num = Num / gcd(Num, Denom);
+  static constexpr uint32_t den = Denom / gcd(Num, Denom);
   using type = ratio<num, den>;
 };
